+++ otx/algorithms/action/adapters/mmaction/task.py
@@ -29,6 +29,7 @@ from mmaction.models import build_model as build_videomodel
 from mmaction.utils import collect_env
 from mmcv.runner import CheckpointLoader, load_checkpoint, wrap_fp16_model
 from mmcv.utils import Config, ConfigDict, ProgressBar, get_git_hash
+from torch import distributed as dist
 
 from otx.algorithms.action.adapters.mmaction import (
     Exporter,
@@ -49,6 +50,7 @@ from otx.algorithms.common.adapters.mmcv.utils.config_utils import (
     update_or_add_custom_hook,
 )
 from otx.algorithms.common.configs.configuration_enums import BatchSizeAdaptType
+from otx.algorithms.common.utils import append_dist_rank_suffix
 from otx.algorithms.common.utils.data import get_dataset
 from otx.algorithms.common.utils.logger import get_logger
 from otx.api.entities.datasets import DatasetEntity
@@ -208,6 +210,7 @@ class MMActionTask(OTXActionTask):
             ckpt = ckpt["model"]
             if not new_path:
                 new_path = ckpt_path[:-3] + "converted.pth"
+            new_path = append_dist_rank_suffix(new_path)
             torch.save(ckpt, new_path)
             return new_path
         return ckpt_path
@@ -234,7 +237,7 @@ class MMActionTask(OTXActionTask):
     def configure_distributed(cfg: Config):
         """Patching for distributed training."""
         if hasattr(cfg, "dist_params") and cfg.dist_params.get("linear_scale_lr", False):
-            new_lr = len(cfg.gpu_ids) * cfg.optimizer.lr
+            new_lr = dist.get_world_size() * cfg.optimizer.lr
             logger.info(
                 f"enabled linear scaling rule to the learning rate. \
                 changed LR from {cfg.optimizer.lr} to {new_lr}"