+++ otx/algorithms/common/utils/__init__.py
@@ -20,6 +20,7 @@ from .callback import (
     TrainingProgressCallback,
 )
 from .data import get_cls_img_indices, get_image, get_old_new_img_indices
+from .dist_utils import append_dist_rank_suffix
 from .ir import embed_ir_model_data
 from .utils import (
     UncopiableDefaultDict,
@@ -42,4 +43,5 @@ __all__ = [
     "get_arg_spec",
     "get_image",
     "set_random_seed",
+    "append_dist_rank_suffix",
 ]