+++ otx/algorithms/detection/adapters/mmdet/models/detectors/__init__.py
@@ -4,6 +4,8 @@
 #
 
 from .custom_atss_detector import CustomATSS
+from .custom_deformable_detr_detector import CustomDeformableDETR
+from .custom_dino_detector import CustomDINO
 from .custom_maskrcnn_detector import CustomMaskRCNN
 from .custom_maskrcnn_tile_optimized import CustomMaskRCNNTileOptimized
 from .custom_single_stage_detector import CustomSingleStageDetector
@@ -16,6 +18,8 @@ from .unbiased_teacher import UnbiasedTeacher
 
 __all__ = [
     "CustomATSS",
+    "CustomDeformableDETR",
+    "CustomDINO",
     "CustomMaskRCNN",
     "CustomSingleStageDetector",
     "CustomTwoStageDetector",