+++ tests/integration/cli/detection/test_detection.py
@@ -68,10 +68,21 @@ default_templates_ids = [default_template.model_template_id]
 templates = Registry("otx/algorithms/detection").filter(task_type="DETECTION").templates
 templates_ids = [template.model_template_id for template in templates]
 
+experimental_templates = [
+    parse_model_template(
+        "otx/algorithms/detection/configs/detection/resnet50_deformable-detr/template_experimental.yaml"
+    ),
+    parse_model_template("otx/algorithms/detection/configs/detection/resnet50_dino/template_experimental.yaml"),
+]
+experimental_template_ids = [template.model_template_id for template in experimental_templates]
+
+templates_w_experimental = templates + experimental_templates
+templates_ids_w_experimental = templates_ids + experimental_template_ids
+
 
 class TestDetectionCLI:
     @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates, ids=templates_ids)
+    @pytest.mark.parametrize("template", templates_w_experimental, ids=templates_ids_w_experimental)
     def test_otx_train(self, template, tmp_dir_path):
         tmp_dir_path = tmp_dir_path / "detection"
         otx_train_testing(template, tmp_dir_path, otx_dir, args)
@@ -90,26 +101,26 @@ class TestDetectionCLI:
         otx_resume_testing(template, tmp_dir_path, otx_dir, args1)
 
     @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates, ids=templates_ids)
+    @pytest.mark.parametrize("template", templates_w_experimental, ids=templates_ids_w_experimental)
     @pytest.mark.parametrize("dump_features", [True, False])
     def test_otx_export(self, template, tmp_dir_path, dump_features):
         tmp_dir_path = tmp_dir_path / "detection"
         otx_export_testing(template, tmp_dir_path, dump_features, check_ir_meta=True)
 
     @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates, ids=templates_ids)
+    @pytest.mark.parametrize("template", templates_w_experimental, ids=templates_ids_w_experimental)
     def test_otx_export_fp16(self, template, tmp_dir_path):
         tmp_dir_path = tmp_dir_path / "detection"
         otx_export_testing(template, tmp_dir_path, half_precision=True)
 
     @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates, ids=templates_ids)
+    @pytest.mark.parametrize("template", templates_w_experimental, ids=templates_ids_w_experimental)
     def test_otx_export_onnx(self, template, tmp_dir_path):
         tmp_dir_path = tmp_dir_path / "detection"
         otx_export_testing(template, tmp_dir_path, half_precision=False, is_onnx=True)
 
     @e2e_pytest_component
-    @pytest.mark.parametrize("template", templates, ids=templates_ids)
+    @pytest.mark.parametrize("template", templates_w_experimental, ids=templates_ids_w_experimental)
     def test_otx_eval(self, template, tmp_dir_path):
         tmp_dir_path = tmp_dir_path / "detection"
         otx_eval_testing(template, tmp_dir_path, otx_dir, args)