+++ tests/unit/algorithms/common/adapters/mmcv/ops/__init__.py
@@ -1,0 +1,4 @@
+"""Test for otx.algorithms.common.adapters.mmcv.ops"""
+# Copyright (C) 2023 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+#