+++ tests/unit/algorithms/detection/adapters/mmdet/test_configurer.py
@@ -7,6 +7,7 @@ from mmcv.utils import ConfigDict
 
 from otx.api.entities.model_template import TaskType
 from otx.algorithms.common.adapters.mmcv.utils.config_utils import MPAConfig
+from otx.algorithms.detection.adapters.mmdet import configurer
 from otx.algorithms.detection.adapters.mmdet.configurer import (
     DetectionConfigurer,
     IncrDetectionConfigurer,
@@ -72,10 +73,14 @@ class TestDetectionConfigurer:
             "torch.distributed.is_initialized",
             return_value=True,
         )
+        world_size = 2
+        mocker.patch.object(configurer, "dist").get_world_size.return_value = world_size
         mocker.patch("os.environ", return_value={"LOCAL_RANK": 2})
         config = copy.deepcopy(self.model_cfg)
+        origin_lr = config.optimizer.lr
         self.configurer.configure_device(config, True)
         assert config.distributed is True
+        assert config.optimizer.lr == pytest.approx(origin_lr * world_size)
 
         mocker.patch(
             "torch.distributed.is_initialized",